Correciones de codigo v2

+++ Padel.Application/DTOs/Productos/CreateProductoDto.cs
@@ -6,5 +6,5 @@ public class CreateProductoDto
 {
     [MaxLength(50)]
     public required string Nombre { get; set; }
-    public required double Precio { get; set; }
+    public required decimal Precio { get; set; }
 }
